+++ BrotatoPrototype/Assets/Scripts/UI/GameMenu/UIWaveResults.cs
@@ -15,7 +15,7 @@ public class UIWaveResults : MonoBehaviour
 
     public void UpdateWaveResults(PlayerCharacteristics playerCharacteristics)
     {
-        int hunger = playerCharacteristics.CurrentHunger;
+        float hunger = playerCharacteristics.CurrentHunger;
         currentHungerTMP.text = hunger.ToString();
 
         int amountOfFoodForWave = playerCharacteristics.GetComponent<PlayerSatiety>().GetAmountOfFoodForWave();
@@ -23,8 +23,8 @@ public class UIWaveResults : MonoBehaviour
 
         amountOfLevelUpsForWaveTMP.text = playerCharacteristics.GetComponent<LevelSystem>().NumberOfLeveledUpForCurrentWave.ToString() + "<sup>х</sup>";
 
-        int currentSatiety = playerCharacteristics.CurrentSatiety;
-        int percentageReductionMaxHealth = 100 - currentSatiety;
+        float currentSatiety = playerCharacteristics.CurrentSatiety;
+        float percentageReductionMaxHealth = 100 - currentSatiety;
         if (percentageReductionMaxHealth < 0) percentageReductionMaxHealth = 0;
 
         percentageReductionMaxHealthTMP.text = "максимальное здоровье \n снижено на " + percentageReductionMaxHealth + "%";
